renamed core to AutoNeuro; minor refactoring
