+++ .gitignore
@@ -1,3 +1,7 @@
-__pycache__/
 *.pyc
-.ipynb_checkouts/
+
+*.egg-info
+__pycache__/
+.ipynb_checkpoints/
+
+results/
